PiSideCode/requirements.txt
git+https://github.com/antmicro/pyrav4l2.git
pyntcore # >= 2023.2.1.2
robotpy-wpimath
Flask
flask[async]
eventlet
opencv-contrib-python >= 4.10.0
flask-socketio
matplotlib
numpy >= 1.26.4
Pillow >= 10.3
cffi
simplejpeg